requirements.txt
faiss-cpu==1.7.1.post2
imutils==0.5.4
install==1.3.4
memoization==0.3.2
numpy==1.21.0
opencv-python==4.5.2.54
psycopg2-binary==2.9.1